requirements.txt
telethon==1.28.5
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
python-socketio==5.9.0
eventlet==0.33.3 